Add bifrost inventory management procedures
The infra-cloud Bifrost inventory needs to have secrets moved into hiera,
and may come pre-populated with random names instead of assigned hostnames in
the infracloud scheme. So we have the -dns tool to name hosts according
to a uniform scheme, and -hiera to move the passwords moved into hiera yaml
files.

Meanwhile some slight tweaks are needed to use this scheme to build
inventory for bifrost, and an initial inventory for hpuswest is checked
in with passwords stripped out. This inventory includes all static IP
assignments for the nodes as well, which should allow us to build DNS
directly from the erb.

Change-Id: I2f0c3f856d3bd5885755670845653d2e001f9081