Add manifest for translate.openstack.org
Manifest for translate.openstack.org and related site.pp change.
This is very similar to the translate_dev.pp we have already.

I've added some application level (wildfly and zanata) versioning
options so we can upgrade the versions on this production server
independent of changes we make in the puppet-zanata module, which
will deploy to the dev server.

Change-Id: I224ab532b31d70aabcec5c7092ce7d4ecbfec9f3