Revert "Add support to launch node for attaching volumes"
This reverts commit f490260de72fe7b155f1ec1db9224e3be250deb7.

We should really use cinder when we're attaching substantial
volumes to servers, and pretty much at this point if we're
attaching volumes to servers, they are substantial.  This amounts
to a feature that, at the moment at least, we should not actually
use, until it can handle cinder.  Remove it so we are not tempted.

Change-Id: Ib0d1031ddebb89086de9bd5b8e6cbeda072a9b59