Turn image logging level up to DEBUG
In I54db78b028017500b58693ec717c23eaf6a220b6 I added some handy values
that are intendend to go into public log files to help you establish
when an image-build has started or stopped (for example, [1] is
currently ~300MiB of looping build log data and it's really hard to
pick out where things are going wrong).

The problem is, I put them in at DEBUG level and the image loggers are
only set to INFO level.  So the messages aren't making it into the
public logfiles as hoped, and thus not available to help people
diagnose issues with their builds.

I think the best idea is to turn these up to DEBUG level to capture
these, and any other low-level messages from this logger, in the
public output.

[1] http://nodepool.openstack.org/image.log

Change-Id: Ib9ac5fc81ccef73f1b1e20b8381bdd1afac4bcae