Update nanliu's git module to puppet-community git
Puppet community's git_resource module will manage the standalone
git provider, as the forked vcsrepo not covers all of our use-case
for git repositories. The development of nanliu's module was moved
under puppet-community. It contains some important fixes related
to commit ref and tag changes, and the mechanisom of notification
of subscribed resources.

Change-Id: I36a2664707f900fe7807d5f05f69423682c3a51a