Update rabbitmq module
There have been no backwards-incompatible changes in the rabbitmq
module since 5.0.0. However there have been major bugfixes to keep up
with changes in upstream rabbitmq packages, so it is worth updating.
This update is needed for Infra Cloud to work.

The only module that uses the rabbitmq module currently is the
storyboard module, all of which is contained in one class[1], so the
impact of this update is minimal.

See the puppetlabs-rabbitmq changelog[2] for more details of what this update entails.

[1] http://git.openstack.org/cgit/openstack-infra/puppet-storyboard/tree/manifests/rabbit.pp
[2] https://github.com/puppetlabs/puppetlabs-rabbitmq/blob/master/CHANGELOG.md

Change-Id: Ief759e6676799864d9ef8b472d30e8f9834b41c0