Use ansible service restart instead of start
The ansible service module is smart and will check the status of a
daemon and only start it if deemed necessary when asked to start a
service. The problem with this is that jenkins can go sideways confusing
the daemon system making it think the service is up even though we have
kill -9'd it into oblivion. The simple work around for this is to use
restart which will always bounce the service in ansible.

Change-Id: I240c318e9fa824d88bb817cf382d50ba2902d668