Update stdlib module to latest version
There are no backwards-incompatible changes between 4.3.2 and 4.9.0.
Upgrading to 4.9.0 gives us new functions such as the
load_module_metadata function, which can be helpful when upgrading
modules since we can have modules inspect what version of a module is
available. See the changelog[1] for more information of what this
update includes.

[1] https://github.com/puppetlabs/puppetlabs-stdlib/blob/4.9.0/CHANGELOG.md

Change-Id: Ia366c0f7f1bfbfa843071e733a5b0a96873a60d8