Clean up references to old pypi mirror urls.
This patch removes old references to pypi.<host>.openstack.org,
the old path aliases from the mirror virtualhost, and adds a new
robots.txt file as well as a symlink to the pypi alias. The last
is mostly a cosmetic change accessible via the filesystem; If clicked
on directly, the mod_alias directive will take over.

Change-Id: If075ab84550165e093247b3890cb13a8c854f8df
Depends-On: I895b419e7755f8ec0e5fa594eebc8f9d79d333ba