Add new statsd variables for hostclouds
Currently there is no variable to aggregate node usage based on
cloud donor. This makes it a challenge to create a graphite graph
to motivate additional cloud resource donatations.

With the goal of trying to publish something like:
http://tinyurl.com/ng56aaj this patch adds a new variable
'hostcloud' which is designed to represent the sum of the
many cloud regions we have captured in the variable 'provider'.

Something will also have to happen in zuul's nodepool.py file but
I don't know what that might be yet.

Change-Id: I72e3edbf7c8e97784e8b6238e6280366ed9f9511