Update the name of repos.yaml to renames.list
In the renaming a project section of the gerrit documentation,
there is a section about running a script against a yaml file.

Currently the documentation directs the reader to call the file
repos.yaml. The script errors on this name, looking for a script
called renames.list. This patch edits the documentation to direct
the reader to create a yaml file called renames.list.

Change-Id: Ia43dd5ae231826331502aa0e5263852fce9cb014