Allow CORS against elasticsearch API
Since we do fine grained access control to the API via a proxy that
limits what requests can be made from anywhere go ahead and allow CORS
requests from anywhere as they won't be able to write data anyways.

This is useful so that you can test a new version of $tool hosted
locally against the actual cluster or for admins to run admin tools
hosted locally against the cluster.

Change-Id: I774d0ad0b246315794ab387acc39c41c7cfac3cd
Depends-On: I0aa8d5167c770c1024b7596da582d6cc089b1b47