Expose enable_unbound in openstack_project::server
The openstack_project::template class has an $enable_unbound parameter
that controls whether the unbound class is declared. Since
openstack_project::server mostly just wraps the
openstack_project::template class, we should be able to pass this
parameter through openstack_project::server into
openstack_project::template.

The use case for allowing more control over unbound is that the
InfraCloud baremetal nodes will use bifrost which sets up a dnsmasq
during the install process. By default dnsmasq listens on localhost:53
which conflicts with unbound listening on the same interface and port.
Currently bifrost doesn't allow us to configure the except-interface
option in dnsmasq.conf, though adding that functionality is probably a
valid alternative to this patch. Since we don't really need unbound
when we control the dhcp server that assigns the domain-name-servers,
it's easiest to just not include it.

Change-Id: I6f7ea252950d78db55411133966eff3b18ef0ec9