Make apply-test aware of modules to upgrade
This patch adds an additional step to apply-test.sh to look for nodes
in site.pp that, in addition to having Node-OS designators, also have a
special designator "Upgrade-Modules". If a node contains that
designator, the script will re-run the install_modules.sh script with
MODULE_FILE set to 'upgrade-modules.env'. This file will contain only
the dangerous modules that we need to upgrade one node at a time.
Rerunning install_modules.sh will have the effect of leaving all
installed modules intact except those listed in upgrade-modules.env.
When we are ready to upgrade a node, we add "Upgrade-Modules" to its
node definition in site.pp to effectively opt-in to using the upgraded
module. The apply-test will avoid testing that node in the first round,
and instead test it in a second round with the updated module.

This patch also fixes the environment variable name of MODULE_FILE (not
MODULE_ENV_FILE), as it needs to be MODULE_FILE for install_modules.sh
to pick it up.

The catch here is that we need our deployment method to also look at
both modules.env and upgrade-modules.env and have logic in place to
merge them if necessary. This will be easier when puppet-apply becomes
a reality. We should not merge this patch until we have that figured
out.

Change-Id: I38143277ec8020d85d1a0687667b6245732a25c3