Logstash 2.0 compat ES output rule
Logstash 2.0 defaults to HTTP elasticsearch output which means that the
elasticsearch output (no _http) does HTTP and a new elasticsearch_java
output exists if you want to continue doing the native api output. We
had been doing HTTP so just need to update the output name. The host
parameter is also deprecated and you must pass an array to the hosts
parameter instead so update that as well.

We had been using a local elasticsearch client daemon to speak http to,
but this uses much more memory leading to OOMing. Reduce the memory by
taking advantage of logstash 2.0's ability to load balance http requests
to elasticsearch itself and stop running an elasticsearch daemon on the
workers.

Note this is not compatbile with Logstash 1.3.3 which is what we are
currently running so this change should only go in as part of an upgrade
to Logstash 2.0 and beyond.

Change-Id: I788ecb936f9fa5a006332ed626f90c33a255d9bf