Identify logstash multiline events by build uuid
We had previously been forced to rely on the source host for the logs
and the log filename to associate events because the last event needed
to be flushed from the logstash pipeline. The only way to do that was to
get a new event in not matching the multiline rule with the same
identity. This was hacky and resulted in confused results if the first
line of the next matching log matched the multiline rule. It would paste
two jobs together essentially.

Thankfully new logstash will timeout events in multiline rules. If they
sit there for 5 seconds (by default) they will be automatically flusehd.
This means we can use a truly unique stream identifier for multiline
parsing. Use build_uuid.filename as the uuid identifies each set of logs
uniquely and the filename is a unique path within that set.

Change-Id: If3f90a79759fc6e5dfc47417ac1f86eee4ed8afe