Allow haproxy to bind to all ports in selinux
By default haproxy can only bind to HTTP(S) ports all other ports can't
be bound due to the selinux policy. Simple fix for this is to toggle the
boolean that allows haproxy to bind any port in the selinux policy. Do
this with an exec that first checks if the boolean is set.

Change-Id: I49c8bdc3586fa82cd954a6ef9be27f48f9a623ec