Manage clouds.yaml on puppetmaster with puppet
For launch_node and for ansible-inventory we need a functioning
clouds.yaml file. The file should really just contain the entries we
actually need, becuase we don't want to be listing all the hosts
in nodepool. However, we do want ALL of the entries available, so the
all-clouds template has them all there. Not sure where we want that file
to go though.

Change-Id: Ibee80f9a4d8f159a626e2b4c2e9639134a559ae4