Allow to disable cron jobs
Any fork of the upstream that is not exposed to frequent
code changes (as the openstack-infra itself), doesn't need to trigger
update every 15 min. This patch add the option to change the "hardcoded"
time interval by disababling the cron jobs. It does not change the way
the upstream is automating its components. It only comes to help any fork
of the upstream to keep track with the upstream with less conflicts.

Change-Id: I23bdd5b7457caa2de9ddbd2a91279bb1f3272207