Separate concerns in pypi_mirror.pp
This patch separates the creation of the bandersnatch pypi mirror
from the creation of the apache vhost, and parameterizes the data
directory in which bandersnatch places its packages.

This is done so that we may reuse the pypi_mirror.pp module when
building our new unified mirrors, which host pypi assets in a
different directory and under a different hostname and URI.

This patch should not trigger any changes on the existing mirrors.

Change-Id: I55286df53d8dd84ea2377035b830cd92d378ad39