Add wheel mirror to mirrors
This patch adds a new "wheel" directory to the pypi mirrors,
as an rsync target for our built wheel packages.

A rewrite rule has been added in anticipation of serving
the wheels from an AFS drive. Since AFS has a practical
folder size limit, we are using /a/a /s/sp/split /s/st/style
directory structure that should be AFS-tolerant. The rewrite
rule creates the necessary mappings that make the packages
available to pip.

Example: HTTP GET /Babel/ -> /B/Ba/Babel/

Furthermore, a cron job has been added to periodically generate
a human-readable index of these mappings, in accordance with
PEP503. While frequently regenerated, this index should only
change meaningfully if a new package is added to the wheel, as
it only represents the package names themselves, rather than
the available versions of said package.

Change-Id: I743fc3ec629eea225c981d6e870751f33e77d7c6