Update subunit job name selection regex
This commit updates the selection regex to be a bit more general an
ensure that the subunit streams from all the gate tempest jobs are
being picked up. There are certain jobs that don't start with
gate-tempest-dsvm or gate-grenade-dsvm that we wanted to collect
results for. For example the puppet jobs will be named something like:

gate-puppet-openstack-integration-scenario001-tempest-dsvm-centos7

and those would not be picked up with the old regex.

Change-Id: If68978679a9ebb6252663691b1850918db913990