Disable Gerrit drafts feature
In Gerrit 2.11 disabling draft pushes in gerrit ACLS was not enough to
disable the drafts feature.  There is now an change.allowDrafts [1]
settting in gerrit.config to disable drafts which also needs to be
set. Gerrit enables drafts by default but we want it disabled
because we prefer using the 'work in progress' (WIP) workflow.

Change-Id: I53f7a80a3b2596a9af32f6f910091fa3af1106e0
depends-on: Iea970a85d433d8707102a3dcaf8cbae2d5662b6e