Upgrade the concat module
The currently installed version of puppetlabs-concat has a bug that
causes concat resources with ensure => absent to be run on every puppet
run. This affects the infracloud module, which uses the
puppetlabs-apache module to configure apache vhosts, which uses the
concat type to ensure the default ssl vhost is absent. This causes the
httpd service to be restarted on every puppet run. The bug was fixed in
1.1.1[1]. This patch upgrades the concat module to 1.2.5, which is the
latest 1.x version.

The concat resource is not used directly by any of our modules[2].
Besides the puppetlabs modules, which we can assume are well-tested
with the latest version of concat, the only other module that appears
to use concat is the selinux module. That module only uses it in the
selinux::restorecond class, which we're not using. So this upgrade
should be safe for us.

[1] https://github.com/puppetlabs/puppetlabs-concat/pull/215
[2] http://codesearch.openstack.org/?q=concat&i=nope&files=%5E.*%5C.pp%24&repos=project-config,puppet-accessbot,puppet-ansible,puppet-apparmor,puppet-apps_site,puppet-askbot,puppet-asterisk,puppet-autossh,puppet-bandersnatch,puppet-bugdaystats,puppet-bup,puppet-cgit,puppet-ciwatch,puppet-dashboard,puppet-diskimage_builder,puppet-drupal,puppet-elastic_recheck,puppet-elasticsearch,puppet-etherpad_lite,puppet-exim,puppet-gerrit,puppet-gerritbot,puppet-github,puppet-grafyaml,puppet-graphite,puppet-haveged,puppet-hound,puppet-httpd,puppet-infra-cookiecutter,puppet-infracloud,puppet-iptables,puppet-jeepyb,puppet-jenkins,puppet-kerberos,puppet-kibana,puppet-lodgeit,puppet-log_processor,puppet-logrotate,puppet-logstash,puppet-mailman,puppet-mediawiki,puppet-meetbot,puppet-mysql_backup,puppet-n1k-vsm,puppet-nodepool,puppet-nslcd,puppet-odsreg,puppet-openafs,puppet-openstack_health,puppet-openstackci,puppet-openstackid,puppet-os_client_config,puppet-packagekit,puppet-pgsql_backup,puppet-phabricator,puppet-pip,puppet-planet,puppet-project_config,puppet-puppet,puppet-redis,puppet-refstack,puppet-releasestatus,puppet-reviewday,puppet-setproxy,puppet-simpleproxy,puppet-snmpd,puppet-ssh,puppet-ssl_cert_check,puppet-stackalytics,puppet-statusbot,puppet-storyboard,puppet-subunit2sql,puppet-sudoers,puppet-surveil,puppet-tmpreaper,puppet-ulimit,puppet-unattended_upgrades,puppet-unbound,puppet-user,puppet-vcsrepo,puppet-vinz,puppet-vitrage,puppet-yum,puppet-zanata,puppet-zuul,system-config

Change-Id: I6abf85fca776cc5cc7f3aee47cc649001e7f15fc