Stop using puppetlabs-mysql for paste
The db for paste.o.o is moving to trove. To do this, we need the paste
class to accept parameters for a remote database. We can also stop
managing the local mysql server and the database, as well as remove the
account_security class which just removes some of the default databases
in the mysql server.

Change-Id: I2046c588193843c5cdae27eb95356a4c4f921feb