Upgrade the stdlib module to 4.10.0
This patch upgrades the puppetlabs-stdlib module from 4.3.2 to 4.10.0.
4.10.0 gives us the load_module_metadata function, which loads the
metadata.json file into scope, plus an improvement to allow the
function to succeed if the metadata.json is not found. With this
function we can read module metadata and use puppet conditionals that
make decisions based on this metadata.

This will help us upgrade the mysql module by allowing us to check
whether metadata.json exists for the mysql module. If it does not
exist, that means the module is "old", and we should use the "old" API.
If it does exist, we can check what version.

This could also help if we want to switch from one module to another
when modules have the same name. We can add conditionals in the puppet
code to check the module author and use the correct interface based on
that.

See the changelog[1] for more information of what this update includes.

[1] https://github.com/puppetlabs/puppetlabs-stdlib/blob/4.10.0/CHANGELOG.md

Change-Id: Ia366c0f7f1bfbfa843071e733a5b0a96873a60d8