Allow the review history panel to resize
Remove the fixed width setting from the review history panel to allow
the comment panel to resize and allow the comments to wrap.  This
will make it easier to read long comments.  This will allow the
history panel to use the entire width of the window and it
will make the comments easier to read, particularly  when there
are a lot of comments.

  preview: https://imgur.com/pIqAu5F

Change-Id: I589c0c000dd8fab9e7f34410a474590a58799bcb