Add a script to list change owner statistics
With the recent upgrade to Gerrit 2.11, changes in its database
schema require updates to our "ATC" generator tools. However, the
REST API in 2.11 is complete enough that it can provide all the
required data, so replace the old MySQL query and SSH API based
tooling which required database access with something simpler to
run.

This new implementation retains the data model of its predecessor
(in other words, it tracks "owners" of merged changes in Gerrit as a
best-effort proxy for identifying contributors who are also
foundation individual members). The new script should be extensible
enough that we can further change this model at a later time, with
consensus from the technical committee. Further, unlike or earlier
method, anyone with a Gerrit account can now generate technical
election rolls, contact lists for Summit registration discount code
mail merges, and so on.

Change-Id: I6df798fc9e7fc4ff59972aff39c2e37f8c28d776
Signed-off-by: Jeremy Stanley <fungi@yuggoth.org>