Add a script to list change owner statistics
With the recent upgrade to Gerrit 2.11, changes in its database
schema require updates to our "ATC" generator tools. However, the
REST API in 2.11 is complete enough that it can provide all the
required data, so replace the old MySQL query and SSH API based
tooling which required administrative access with something anyone
can run. For the first time, this makes it possible for anyone with
a Gerrit account to generate technical election rolls, contact lists
for Summit registration discount code mail merges, and so on.

Change-Id: I6df798fc9e7fc4ff59972aff39c2e37f8c28d776