Allow openstack_citest 'grant option' privilege
The old version of the puppet mysql module defined the "all" privilege
as all the columns in the *_priv rows in the "db" table of the "mysql"
database, which included "Grant_priv". The new version simply uses the
normal MySQL API, which defines the "all" privilege as "Grant all
privileges at specified access level except GRANT OPTION".

In order to get back the old behavior of grant permissions on single use
slaves we exec to GRANT ALL WITH GRANT OPTION for the test user if the
flag to get all options is passed in. This allows us to continue to rely
on using ALL rather than writing out a list of options that may or may
not change over time.

Change-Id: I93b9375de916ca701399aee03dfaa9ab4d197319