gerrit: Restore reasonable max-width to change info table
Info column width limit got changed to 100% to allow more room for CI table.

This previous change impacted the change info table which now takes
the full page width whenever possible. This makes it difficult to read and use.
The options and buttons are now located to the far right of the screen while
info are left aligned and located in the middle of the screen. The distance
between the info and options/buttons is just too big.

This change restore a reasonable max-width to change info table.
CI table will still be able to expand to full page width.

Change-Id: I04ce684cdb78317802c711cd806573b7f9353cbc