Add AFS apt mirror
The reprepro class in this is in-tree rather than in its own module
purely for ease of getting started. It's also highly hard-coded rather
than flexible.

This change will need a mirror.apt volume and service/reprepro principal
and keytab to be created before it lands.

Local testing indicates that a trusty mirror is 86G.

Change-Id: I84f6a0391f80e6bf567c4bfc18a41bd270fe8c01