Add AFS apt mirror
The reprepro class in this is in-tree rather than in its own module
purely for ease of getting started. It's also highly hard-coded rather
than flexible.

This change will need a mirror.apt volume and service/reprepro principal
and keytab to be created before it lands.

Allow for pool trimming after a 2 hour delay. Each devstack run of
apt-get update should be able to be assumed to be valid for the length
of the devstack. For that reason, only delete files that are
unreferenced during the subsequent mirror run, ensuring at least a 2
hour delay between becoming unreferenced and going away.

Local testing indicates that a trusty mirror is 86G.

Change-Id: If13498eae3872fbfdd5a185088395803095bbf88