Don't exclude python-setuptools
The centos-minimal dib build is failing as it tries to cache some
packages that have dependencies on python-setuptools.

Since I556b2cec249ef46e09ffca3cd75521e0beeb7779 we have been
installing a yum.conf that has an "excludes" for this package.

We have not noticed this on the existing dib builds because they have
python-setuptools already installed in the base-image as part of the
cloud-init install.  So this satisfies the dependency; however on the
minimal builds the package is not there, and the exclude means the
packages dependencies that can not be satisfied and the build stops.

The original code has since been removed.  I think the problem of some
parts of the packaged setuptools conflicting was probably due to the
large version delta between whatever was in CentOS6 to a current pip
install.  Clearly the issue doesn't exist with pip installing over the
Centos7 version (because it's working now), but for an abundance of
caution let's pre-install the packaged version, clear it out, then
install from pip.

Change-Id: I35408717dc340271dea3d857bc19ea1590e84361