Added afs to wheel mirror slaves
This removes the SSH keys from the wheel mirror slaves, and
replaces them with the AFS share. It does not yet add afs
credentials. As we have not yet provisioned our wheel slaves,
no manual updates of the servers should be required.

Hiera keys for the previous approach should be removed.

Change-Id: Ifebf5d53d80e934674704078f7dd675f77aef5aa