Added AFS keytab to wheel slave
This adds a hiera-loaded keytab for the wheel slaves, keyed off
off the node's distid and version. It may be used when writing
wheels to the AFS share directory.

Note: This will require the creation of a key per wheel mirror,
as well as the appropriate folder and access rights in AFS.

Change-Id: I81c39d420d8ac70def57949ea0d4c323b8797086