Use diskimage-builder from git for nodepool
Its more-or-less reality that nodepool and diskimage-builder are very
interrelated and nodepool is a defacto last-line of CI for
diskimage-builder.  If a change to dib has passed all CI there, then
there will not really be any further testing until it hits nodepool
the next time it releases.  So instead of adding this latency and the
consequent batching of many changes, let's install dib from the latest
git.

The risk here is low because the daily image generation is quite
tolerant to broken builds, should they occur.  We are better off
fixing issues that occur as soon as possible, and people are watching
the nodepool builds fairly carefully.  If a change does introduce an
issue, then we can submit the change through dib and rebuild, helping
our ability to iterate.

Change-Id: I6bd61b6b7be8a71f7052df6b966e8e6af0384d9d