Temporarily stop releasing after bandersnatch
We're in the middle of a long initial vos release to a new RO site
and we don't want the bandersnatch cron to accidentally start a new
one when this finishes.  Instead, we want to manually release until
we are certain that the release runtime is down to a manageable
level.

Change-Id: Ib5af8d3bbd4439fc414e6df87e6f8b37807fa138