Add omfracloud to nodepool
This adds the omfracloud jenkins account credentials to nodepool.
I'm not pleased with the file resource in the node definition, but that
node definiton grew huge and needs a refactor anyways so we can do that
when we do it.

I have verified that the correct keys are in hiera.

Change-Id: Iafca5e86f72321c6aa7bef748ac2b1942539d15f