Cleanup of nodepool builder logging
We can use textwrap to break up the ridiculously long lines in the
logging configuration (it's less ridiculous than it used to be, but
still).  Only trick is to put spaces between the arguments to give it
somewhere to split (and better for readability) and stop it splitting
on words/hyphens.

Since the last run, the old devstack-centos7-dib has gone too
(I99b8607a6cd7347614118005dbe89ee46ec4a772)

Change-Id: I04840e98b0564dc531b74b96e8595f9d18b284af