Cleanup of nodepool builder logging
We can use textwrap to break up the ridiculously long lines in the
logging configuration (with less images, it's less ridiculous than it
used to be, but it's still easier to parse).  Only trick is to put
spaces between the arguments to give it somewhere to split (and better
for readability) and stop it splitting on words/hyphens.

Change-Id: I04840e98b0564dc531b74b96e8595f9d18b284af