Update paths for AFS mirror
In the current setup, the reprepro db and lists dirs are being written
into /root. Move them into AFS so that we store all the data in the same
place.

Also, update the script to point to the confdir for reprepro commands.

Note: This will require a manual move of the current db and lists dirs
from /root to /afs/.openstack.org/mirror/apt before this is run.

Change-Id: I2638621e6af98296704e680227fae174d8a0e541