Update InfraCloud SSL configuration
Update the system-config manifest to support the simplifying changes
made in the puppet-infracloud module.

This patch will require updates to hiera. We need keys
ssl_cert_file_contents and ssl_key_file_contents added, and
ssl_cert_file_contents must be in the 'infracloud' hiera group since it
is shared to the compute nodes.

Change-Id: I39c70b1077e8b467e0a7e123a694d037ffc77f7a
Depends-On: Ibeea608e965e58c496a95b2f02a4bf6b13e15f0e