Add support to shade-launch-node for cinder attach
Now that we have a shade version of the launch node script adding in
support for attaching a cinder volume is simple. Do this so that
launching mirrors which rely on cinder volumes is simpler.

This updates the mount_volume.sh script to setup the first cinder volume
with lvm and mount it under the specified path.

Change-Id: Ie95fd4bd5fca8df4f8046d43d1333935cad567e3