Send ansible-puppet timing information to graphite
This should generate two graphs. One will be a graph that cycles between
0 and 1. The 1 state is when ansible-puppet is applying and the 0 state
is when puppet ansible-puppet is idle. A second graph will be the timing
of how long it took. This will rise to the time of how long it took,
then stay there until a new value is set. This is because of limitations
in statsd gauging. Still, an average() function over the second graph
will show us what the average time to converge is for our ansible-puppet
rube goldberg machine.

Change-Id: I74f19e038ba5b4576cba5ce9ce3f1dd64e5a44cb