Infra-cloud-west is currently offline
The infra-cloud-west is currently unavailable due to being relocated
causing the inventory for ansible to not be generated:
Inventory script (/etc/ansible/hosts/openstack) had an execution error: Error fetching server list on openstackci-infracloud-west:RegionOne:

This means ansible and hence puppet fail to run.

For now remove infra-cloud-west from clouds.yaml. This should be added back
in one the cloud is turned back on.

Change-Id: I16fe54c3c040c273977b0cb2e51bf7f804832380