Customize cgit max-repo-count and set it to 1500
The max-repo-count setting in the cgitrc adjusts how many repos should
be displayed on each page. In the puppet-cgit module we have a default
of 600, which was intended to be enough for all the repos to be shown
on git.openstack.org without a second page. We're now at over 1,000
repos, so you now need to page through them and this is not an optimal
experience. Since cgit can handle loading thousands of repos on the
main page, add the setting to our local configuration and bump it up
to 1500.

Change-Id: I7a2ea185f0876806ecfd17bedd84dbcd9769b124