Initial commit for RPM afs mirror
This is the first step to getting an AFS mirror for centos going. We
still need to add RPM validation and enabled automatic vos release.

However, this will atleast enabled us to do the initial import of the
RPMs into AFS which will take a few days to mirror properly.

Change-Id: Ic3b0ca5f5c9321b7e5048f4000ca58266d117083
Signed-off-by: Paul Belanger <pabelanger@redhat.com>