Don't install ES on logstash workers
Now that we are running logstash 2.0 the logstash daemon can talk to
elasticsearch directly and load balance across the cluster. This means
we don't need a local elasticsearch daemon to do that for us. The big
savings here is in memory so stop installing and running elasticsearch
completely on the workers.

Note this will not uninstall an existing ES install you will need to
clean that up if a preexisting install is present.

Change-Id: I9b622674a74a26e7c3024e684e05291f43aec021