Disable automatic installation of backports
In Ubuntu, the archive configures backports as something you must
explicitly request when installing packages. One does this by specifying
a package name of 'backports/$packagename'. But our reprepro mirrors of
backports do not have this flag set, and so, all systems get backported
versions of packages, instead of the one from the released stable
archive.

Change-Id: I8f266dc7045e971567ed934d1206c7b83cacd178