Fix tab name indices in status template
When the extra OpenStack Health link was removed in
I293e326e45e1939dc57556d1477434cd1cfb56a2, the indices for the
'tabsName' array were not updated to reflect the change causing all
links after 'Rechecks' to be off by one. This patch restores the
correct index order to fix the issue.

Change-Id: I4eea976c3fb1a0b9b75d38ba5895b7d5fa5b0de3