Tell os-client-config that tripleo doesn't have neutron
The TripleO cloud does not work when the neutron-based floating ip
logic creates a floating ip directly on a server port. The problem is
that the floating IP is correctly created and correctly associated with
the server, but Nova has no clue that this has happened, so GET calls to
/server do not show that the server has the floating IP. As that's the
way we track success, we can't use it. However, if a floating IP is
created using the nova proxy, it all works - so just configure the cloud
with "has_network: False" so that direct neutron interactions are not
attempted.

Change-Id: Idb90cad36feb99bc5aa19da4874ce97f868681c7