Apt mirror for MariaDB
This patch adds a mirror for the MariaDB trusty packages to aid
the process of making tests involving MariaDB more robust. The
patch uses reprepo as with all other mirrors implemented in
OpenStack-CI.

Note:
The Rackspace mirror is used as previous experience has informed
us that the osuosl mirror has rate limiting which causes mirror
attempt failures, whereas the Rackspace mirror does not.

Change-Id: Ic549bbbd4f5dafc6555174f3ef3bd713b9a727d4