Added clean exit hooks to npm registry script
This patch adds the openstack-registry-hooks project, which will
check the current update sequence of the mirror after every package
has been downloaded, and exit the process cleanly if the update
sequence matches the latest sequence from the registry.

This will effectively de-daemonize the registry-static script,
assuming that it can update in a reasonable amount of time.

Note: When this lands, the index of the registry should be cleaned,
and the entire index should be rebuilt. This should not impact the
tarballs, only the json files, as shasums are checked before
downloading.


Change-Id: Ia94187f1f1d60ad2cbb3e71f76502bec75e228ae