Start using read/write AFS volume for PyPI mirror
The vos release on AFS for PyPI is in a bad place emotionally right now.
So as not to stress it out too much more, stop vos releasing until we
can get it fixed. Also, point the AFS client mirrors to the read/write
path instead of the read only path so that the gate can get PyPI updates
again.

Change-Id: I5d9ac9cc56c16dacad22c1da157cad7f5480c301