Work around bandersnatch not supporting pep503
Temporary work around so our mirrors work with the latest pip release
that expects .'s to be replaced with -'s.

This is crude but considered easier than pinning pip across our services.
It may miss some edge cases so we'll see how it goes.

This should be reverted once bandersnatch supports pep503:
https://bitbucket.org/pypa/bandersnatch/pull-requests/20/fully-implement-pep-503-normalization/diff

Change-Id: Ic4acf4c1d4e78d2524ea49356c61916947fd9a3d