Stop using data in openstack_project class
For reasons, a class called simply 'openstack_project' has had some data
stored in it. Randomly throughout our code, this data is grabbed at.
Hiera is the correct way to grab data like this.

Typically we have always done all hiera lookups in site.pp, then passed
variables around to get the data where we need it. This approach is
implemented in this patch.

We could also (and I tend to think this is the correct move) break with
the past and start allowing hiera() function calls from anywhere in
system-config, especially for data stored in the public data hiera tree.
This mirrors what we were already doing in system-config with the
openstack_project::variablename stuff anyways.

Change-Id: I46b7312ea07eb6fb3e5b77c7c70d788ab8a9d35d