Have zuul use the copied jenkins jobs directory
We have puppet copy the contents of the jenkins_jobs directory
in the project-config repo to a new directory in /etc, principally
so that puppet can detect when files in that directory have changed
(as opposed to when anything in the repo changes) and perform a
reload.

Tell zuul to use those copied files.  This avoids a race where the
vcsrepo may update after the file copy, which would delay the zuul
update by a puppet-ansible pulse.

Change-Id: I4dfd74c7d72dcf9a920c55143d56329ada7186ab