Do not fail ansible if one cloud is down
When a cloud isn't available for whatever reason (wrong credentials or an
outage etc) the default behavour when building an ansible inventory of
hosts is to bail. Instead, we want to still be able to run ansible on
any hosts that we can still access. Set this new config variable to do so.

The ansible feature just merged here:
https://github.com/ansible/ansible/pull/14699

We pull in the latest version of ansible so we should be safe to merge this.

Change-Id: I536efb130965bf6a1b31de2f454ad4985b54edd8