Bump PuppetDB version to 4.1.0 to get ssl related params
The ability to specify the paths for ssl certificates and keys
was introduced on version 4.1.0:

https://github.com/puppetlabs/puppetlabs-puppetdb/tree/4.1.0

The current version doesn't expose those params, forcing a user
to set that configuration by hand on jetty.ini.
This is in preparation of splitting the openstack_project::puppetdb
class and put it on puppet_openstackci.

Change-Id: I85e41e65bdca30dae3767cfacf4fc11100fdc63e