Logstash matching for anisble console output
Unlike Jenkins console log, ansible outputs a different timestamp with
no delimitation (|) before the message.

This means currently we've got a _grokparsefailure for console logs
and they are all going into the one message (actually, TBH I think we
should probably be *putting* these all in the one message, I'm not
sure logging individual lines is actually that helpful.  But anyway).

The ansible lines just look like

 2016-06-20 02:48:09.190192 [Zuul] Launched by zlstatic01

So modify the match for an optional deliminator to the grok for this.
TIMESTAMP_ISO8601 matches this timestamp (tested with grokdebug).

Also, modify the multiline pattern to not look for the deliminator.  I
think the basic assumption "line without timestamp is part of previous
message" is sufficient for all cases.

Change-Id: I16d3215c94f0cd5698fc23b1f1bf6252d9157246