Add zuul user for launching in puppet
Currently we add a zuul user to images during the build with the
"devuser" dib element (see I0e76931fdb4ca0c7445b1e72dc348f0cf03eaee5).

This is not really a great way to add the user.  Firstly, devuser only
adds a single user; so if people are creating a test-image and add a
devuser for their own login purposes, the zuul-worker element
overwrites that to make the zuul user.  Also, this element by default
installs the .authorized_keys to the host from the current build-users
.authorized_keys (assuming that you're in a development environment,
and you're building this image for yourself and you will log in
later).  That's why we had to add the public-key to nodepool in
Idb4ef11576671180060fb5b2b1202f9bfec5fd47 -- but I don't think the
list of public-keys on nodepool is necessarily what we actually want
for the zuul user.

It seems like a much better idea to make this user in the slave puppet
along with where we create all other users.  This adds the zuul user
to slave_common.pp and adds in templates for it's own authorized_keys
file (seeded just from jenkins) and separate sudo template like
jenkins user.

Change-Id: I7d72e3474b1e14b7e4ee91bcc60ad3188d1a6f7e