Revert "Add election alias to governance.openstack.org"
This reverts commit 91dcb510bace2f3a14745c18e945f8504a325c15.

The implementation ignored that the referenced template is being
used by many vhosts but the lists for aliases and directories were
defined at (and referenced from) the containing class scope. This
coupled with the newly introduced conditional for the directories
array caused all other vhosts using this template to no longer get
Directory blocks for their docroots.

Change-Id: I808e654a6fb77440e7aecbde4456ddc720fe0d9a