Move reprepro bits from release to propose slave
When I initial picked release.slave.o.o I thought we needed to run
reprepro on the same server we did vos release on.  This was not
correct, as seen by our wheel servers.

The reason for moving it now, also give us the ability to work around
the wheel-build mutexes, since we are actually doing a `vos release`
ATM, it will allow for faster debugging in the experimental pipeline.

I'll be manually deleting the reprepro bits from release.slave.o.o.

Change-Id: Ic7311b3195ad6f2e2a82a80aa3bf9c755dd99c98
Signed-off-by: Paul Belanger <pabelanger@redhat.com>