Use the Release Bot account on the signing node
The signing node will be used, among other things, for signing and
pushing Git tags as a part of our new release automation. To this
end, a unique Gerrit account "OpenStack Release Bot" is used with an
E-mail address matching that of the OpenPGP signing key in the GnuPG
secring.

Change-Id: Ibf203b71226a672f707e87f22cbf4990b2ebb003
Depends-On: Id07bd1e9f08f2d7c7ea6653867d7440a8ade6ee6