Add ntp-perl package for RedHat platforms
Since I29a101eb9ea5c2d0da6b7762fe93366f0c0a8999 we use ntp-wait, which
is part of the ntp-perl package on RedHat platforms, but included in
the base ntp package for others.  For consistency, add the ntp-perl
package here so we can assume the utility is everywhere.

I have submitted similar upstream (see Jira issue); we could revert
this if it gets accepted.

Change-Id: If59acfb2a950f74d8246424580be3b2bd1409ad3