Add Debian Percona mirror
Kolla requires a few mirrors in order to operate without
depending on external mirrors that can go up and down.  This
work is a copy of Jessie's work on ubuntu-cloud-archive with
alot of search and replace.

I am not super confident in the work, but have reviewed it for
errors.  I'd appreciate a thourough review from folks that
understand how this works.

Change-Id: I01f2287201cbf57f71bceb015260ea1ba91f4ae4