Use pygments for syntax highlighting for git.o.o
One of the options for doing syntax highlighting with cgit is pygments.
Pygments is, itself, rather full featured and in use in plenty of
places. Let's use it.

In order to do this 'sanely' - provide a wrapper script which executes
the highlighting code from python2 and also that does not output inline
CSS. Instead, put that CSS into our static CSS file.

Change-Id: Ia3dc05c541512882a24eaa6bed1a478e2c621f09