Enable ntpdate service for centos-7
We need to ensure ntpdate service starts on boot for centos-7.
Currently, ntpd explicitly require ntpdate to be running before
the sync process can happen in ntpd.  As a result, if ntpdate is not
running, ntpd will start but fail to sync because of DNS cannot be
resolved.

If the clock is not already synced when ntpd starts, it will enter a
synchronisation phase for a default of 900s that holds up ntp-wait.

Change-Id: I98029a288c9a57f4b4b278b0dfb185609989662d
Signed-off-by: Paul Belanger <pabelanger@redhat.com>