Add openstackzuul-tripleo-test-cloud-rh1 cloud config
Today, tripleo-test-cloud-rh1 is running some private infrastructure
in the same project that nodepool is currently running. This is a
problem because is create additional warning message in our logs about
unknown servers.

I've obtained a new user / project which we'll eventually with
nodepool. But first, lets add the credentials so we can run cloud
launcher.

Change-Id: Ic87a9f3f7eccb44396d3e1f30ced9d6e70d186a9
Signed-off-by: Paul Belanger <pabelanger@redhat.com>