Also mirror source packages for debian / ubuntu
As the deb-packaging team moves forward, they are using our AFS
mirrors more and more.  The request has been made to also mirror
package source files which are needed for packaging tasks.

Before this is merged, we'll need to manually hold the locks on both
debian / ubuntu mirrors and manually sync the packages.  This is
because our 30 minute timeout won't be long enough to do this first
task.

Change-Id: I52761116b651402f34349dd895587804cedad7c0
Signed-off-by: Paul Belanger <pabelanger@redhat.com>