Move snmpd module from template class to server
Some jobs running on single-use nodes are upset by the presence of a
running snmpd, and we don't actually query those anyway so it's just
dead weight there. Move the snmpd Puppet module inclusion from the
template class (which is instantiated by the single_use_slave class)
to the server class (which is not, but is still instantiated by the
general slave class we use for persistent job nodes as well as by
our control-plane servers).

Change-Id: Iaee7128a872353cb548d2f55ae36312efd295ea7