Stop triggering offline Gerrit reindexing
Now that Ie88b7effda08996f7c0f325543be703bcf09c794 from
openstack-infra/gerrit is in use (backport of the
https://gerrit-review.googlesource.com/77890 upstream change), we no
longer need to reindex offline. What's more, we've been unable to
do it under our Puppet timeout for ages, so this has been more of a
liability for us than a benefit anyway. As such, unset the
secondary_index option so the offline reindex exec will no longer be
triggered on upgrade.

Change-Id: Ie8f248a7d5fc9f4ef5b64b7d18a99c75393b1b20