Temporarily add rabbit keys to hiera
Rabbit has been changing keys recently, first because the signing
key was compromised, and another time to add better security.
However, puppet module has not been updated since May, and is
not reflecting the proper keys, causing errors when installing
rabbitmq-server.
As a workaround, set the proper keys with hiera until we wait
for a new version of rabbitmq package, to allow us to deploy
infracloud successfully.

Change-Id: I9b449b5633b02d6f43b3f4a085010f5356dfb23e